X: Fix crash bug + other changes from this machine

+++ X/Independent/Core/State/X_State_GameOver.cpp
@@ -57,7 +57,7 @@ void X_State_GameOver::Initialise()
     const float fBaseX = GLToy_Render::GetMinX() + 0.05f;
     GLToy_UI_System::CreateLabel( "G A M E   O V E R", fBaseX, 0.85f );
     GLToy_String szScore;
-    szScore.SetToFormatString( "You scored %d", X_Score::GetScore() );
+    szScore.SetToFormatString( "You scored %d", static_cast< int >( X_Score::GetScore() ) );
     GLToy_UI_System::CreateLabel( szScore, fBaseX, 0.75f );
     GLToy_UI_System::CreateLabel( "Click to continue", fBaseX, 0.65f );
 }
